Replace createSecurePair with TLSSocket + TCP server
This is stupid. Not how to properly replace createSecurePair. SecurePair
fires ‘secure’ event synchronously. Using a couple of sockets instead of
it means that the flow becomes more asynchronous. Therefore socket ‘end’
can not be used directly to control the Connection state, the data first
needs to pass through the TCP server and TLSSocket.